refactor timestamp related tests in modules
