+++ test-framework/sudo-compliance-tests/src/timestamp.rs
@@ -4,6 +4,10 @@ use sudo_test::{Command, Env, User};
 
 use crate::{Result, PASSWORD, SUDO_RS_IS_UNSTABLE, USERNAME};
 
+mod remove;
+mod reset;
+mod validate;
+
 #[test]
 fn credential_caching_works() -> Result<()> {
     let env = Env(format!("{USERNAME} ALL=(ALL:ALL) ALL"))
@@ -16,149 +20,6 @@ fn credential_caching_works() -> Result<()> {
         .as_user(USERNAME)
         .exec(&env)?
         .assert_success()
-}
-
-#[test]
-#[ignore = "gh394"]
-fn sudoers_defaults_timestamp_timeout_nonzero() -> Result<()> {
-    let env = Env(format!(
-        "{USERNAME} ALL=(ALL:ALL) ALL
-Defaults timestamp_timeout=0.1"
-    ))
-    .user(User(USERNAME).password(PASSWORD))
-    .build()?;
-
-    // input valid credentials
-    // wait until they expire / timeout
-    // try to sudo without a password
-    let output = Command::new("sh")
-        .arg("-c")
-        .arg(format!(
-            "echo {PASSWORD} | sudo -S true; sleep 10; sudo true"
-        ))
-        .as_user(USERNAME)
-        .exec(&env)?;
-
-    assert!(!output.status().success());
-    assert_eq!(Some(1), output.status().code());
-
-    let diagnostic = if sudo_test::is_original_sudo() {
-        "a password is required"
-    } else {
-        "incorrect authentication attempt"
-    };
-    assert_contains!(output.stderr(), diagnostic);
-
-    Ok(())
-}
-
-#[test]
-#[ignore = "gh394"]
-fn sudoers_defaults_timestamp_timeout_zero_always_prompts_for_password() -> Result<()> {
-    let env = Env(format!(
-        "{USERNAME} ALL=(ALL:ALL) ALL
-Defaults timestamp_timeout=0"
-    ))
-    .user(User(USERNAME).password(PASSWORD))
-    .build()?;
-
-    // input valid credentials
-    // try to sudo without a password
-    let output = Command::new("sh")
-        .arg("-c")
-        .arg(format!("echo {PASSWORD} | sudo -S true; sudo true"))
-        .as_user(USERNAME)
-        .exec(&env)?;
-
-    assert!(!output.status().success());
-    assert_eq!(Some(1), output.status().code());
-
-    let diagnostic = if sudo_test::is_original_sudo() {
-        "a password is required"
-    } else {
-        "incorrect authentication attempt"
-    };
-    assert_contains!(output.stderr(), diagnostic);
-
-    Ok(())
-}
-
-#[test]
-fn flag_reset_timestamp() -> Result<()> {
-    let env = Env(format!("{USERNAME} ALL=(ALL:ALL) ALL"))
-        .user(User(USERNAME).password(PASSWORD))
-        .build()?;
-
-    // input valid credentials
-    // invalidate them
-    // try to sudo without a password
-    let output = Command::new("sh")
-        .arg("-c")
-        .arg(format!(
-            "echo {PASSWORD} | sudo -S true; sudo -k; sudo true"
-        ))
-        .as_user(USERNAME)
-        .exec(&env)?;
-
-    assert!(!output.status().success());
-    assert_eq!(Some(1), output.status().code());
-
-    let diagnostic = if sudo_test::is_original_sudo() {
-        "a password is required"
-    } else {
-        "incorrect authentication attempt"
-    };
-    assert_contains!(output.stderr(), diagnostic);
-
-    Ok(())
-}
-
-#[test]
-#[ignore = "gh395"]
-fn flag_validate_revalidation() -> Result<()> {
-    let env = Env(format!(
-        "{USERNAME} ALL=(ALL:ALL) ALL
-Defaults timestamp_timeout=0.1"
-    ))
-    .user(User(USERNAME).password(PASSWORD))
-    .build()?;
-
-    // input valid credentials
-    // revalidate credentials a few times
-    // sudo without a password, using re-validated credentials
-    Command::new("sh")
-        .arg("-c")
-        .arg(format!(
-            "set -e; echo {PASSWORD} | sudo -S true; for i in $(seq 1 5); do sleep 3; sudo -v; done; sudo true"
-        ))
-        .as_user(USERNAME)
-        .exec(&env)?
-        .assert_success()
-}
-
-#[test]
-#[ignore = "gh395"]
-fn flag_validate_prompts_for_password() -> Result<()> {
-    let env = Env(format!("{USERNAME} ALL=(ALL:ALL) ALL"))
-        .user(User(USERNAME).password(PASSWORD))
-        .build()?;
-
-    let output = Command::new("sudo")
-        .arg("-v")
-        .as_user(USERNAME)
-        .exec(&env)?;
-
-    assert!(!output.status().success());
-    assert_eq!(Some(1), output.status().code());
-
-    let diagnostic = if sudo_test::is_original_sudo() {
-        "a password is required"
-    } else {
-        "incorrect authentication attempt"
-    };
-    assert_contains!(output.stderr(), diagnostic);
-
-    Ok(())
 }
 
 #[test]
@@ -186,122 +47,6 @@ fn by_default_credential_caching_is_local() -> Result<()> {
         "a password is required"
     } else {
         "incorrect authentication attempt"
-    };
-    assert_contains!(output.stderr(), diagnostic);
-
-    Ok(())
-}
-
-#[test]
-fn flag_reset_timestamp_has_a_local_effect() -> Result<()> {
-    let env = Env(format!("{USERNAME} ALL=(ALL:ALL) ALL"))
-        .user(User(USERNAME).password(PASSWORD))
-        .build()?;
-
-    let child = Command::new("sh")
-        .arg("-c")
-        .arg(format!(
-            "set -e; echo {PASSWORD} | sudo -S true; touch /tmp/barrier1; until [ -f /tmp/barrier2 ]; do sleep 1; done; sudo true"
-        ))
-        .as_user(USERNAME)
-        .spawn(&env)?;
-
-    Command::new("sh")
-        .arg("-c")
-        .arg("until [ -f /tmp/barrier1 ]; do sleep 1; done; sudo -k; touch /tmp/barrier2")
-        .as_user(USERNAME)
-        .exec(&env)?
-        .assert_success()?;
-
-    child.wait()?.assert_success()
-}
-
-#[test]
-fn flag_remove_timestamp_has_a_user_global_effect() -> Result<()> {
-    let env = Env(format!("{USERNAME} ALL=(ALL:ALL) ALL"))
-        .user(User(USERNAME).password(PASSWORD))
-        .build()?;
-
-    let child = Command::new("sh")
-        .arg("-c")
-        .arg(format!(
-            "echo {PASSWORD} | sudo -S true; touch /tmp/barrier1; until [ -f /tmp/barrier2 ]; do sleep 1; done; echo | sudo -S true"
-        ))
-        .as_user(USERNAME)
-        .spawn(&env)?;
-
-    Command::new("sh")
-        .arg("-c")
-        .arg("until [ -f /tmp/barrier1 ]; do sleep 1; done; sudo -K && touch /tmp/barrier2")
-        .as_user(USERNAME)
-        .exec(&env)?
-        .assert_success()?;
-
-    let output = child.wait()?;
-
-    assert!(!output.status().success());
-    assert_eq!(Some(1), output.status().code());
-
-    let diagnostic = if sudo_test::is_original_sudo() {
-        "1 incorrect password attempt"
-    } else {
-        "incorrect authentication attempt"
-    };
-    assert_contains!(output.stderr(), diagnostic);
-
-    Ok(())
-}
-
-#[test]
-fn effect_flag_remove_timestamp_is_limited_to_a_single_user() -> Result<()> {
-    let second_user = "ghost";
-    let env = Env("ALL ALL=(ALL:ALL) ALL")
-        .user(User(USERNAME).password(PASSWORD))
-        .user(User(second_user).password(PASSWORD))
-        .build()?;
-
-    let child = Command::new("sh")
-        .arg("-c")
-        .arg(format!(
-            "echo {PASSWORD} | sudo -S true; touch /tmp/barrier1; until [ -f /tmp/barrier2 ]; do sleep 1; done; sudo -S true"
-        ))
-        .as_user(USERNAME)
-        .spawn(&env)?;
-
-    Command::new("sh")
-        .arg("-c")
-        .arg("until [ -f /tmp/barrier1 ]; do sleep 1; done; sudo -K && touch /tmp/barrier2")
-        .as_user(second_user)
-        .exec(&env)?
-        .assert_success()?;
-
-    child.wait()?.assert_success()
-}
-
-#[test]
-fn flag_reset_timestamp_also_works_locally() -> Result<()> {
-    let env = Env(format!("{USERNAME} ALL=(ALL:ALL) ALL"))
-        .user(User(USERNAME).password(PASSWORD))
-        .build()?;
-
-    // input valid credentials
-    // invalidate them
-    // try to sudo without a password
-    let output = Command::new("sh")
-        .arg("-c")
-        .arg(format!(
-            "echo {PASSWORD} | sudo -S true; sudo -K; sudo true"
-        ))
-        .as_user(USERNAME)
-        .exec(&env)?;
-
-    assert!(!output.status().success());
-    assert_eq!(Some(1), output.status().code());
-
-    let diagnostic = if sudo_test::is_original_sudo() {
-        "a password is required"
-    } else {
-        "Authentication failed"
     };
     assert_contains!(output.stderr(), diagnostic);
 
@@ -438,55 +183,4 @@ fn cached_credential_shared_with_target_user_that_is_self() -> Result<()> {
     }
 
     panic!("timed out")
-}
-
-#[test]
-fn flag_reset_timestamp_plus_command_prompts_for_password() -> Result<()> {
-    let env = Env(format!("{USERNAME} ALL=(ALL:ALL) ALL"))
-        .user(User(USERNAME).password(PASSWORD))
-        .build()?;
-
-    let output = Command::new("sh")
-        .arg("-c")
-        .arg(format!("echo {PASSWORD} | sudo -S true; sudo -k true"))
-        .as_user(USERNAME)
-        .exec(&env)?;
-
-    assert!(!output.status().success());
-    assert_eq!(Some(1), output.status().code());
-
-    let diagnostic = if sudo_test::is_original_sudo() {
-        "a password is required"
-    } else {
-        "incorrect authentication attempt"
-    };
-    assert_contains!(output.stderr(), diagnostic);
-
-    Ok(())
-}
-
-#[test]
-fn flag_reset_timestamp_plus_command_does_not_invalidate_credential_cache() -> Result<()> {
-    let env = Env(format!("{USERNAME} ALL=(ALL:ALL) ALL"))
-        .user(User(USERNAME).password(PASSWORD))
-        .build()?;
-
-    // the first command, `sudo -S true`, succeeds and caches credentials
-    //
-    // the second command, `sudo -k true`, prompts for a password and fails because no password is
-    // provided
-    //
-    // the last command, `sudo true`, is expected to work because `sudo -k true` did *not*
-    // invalidate the credentials
-    //
-    // the `[ $? -eq 1 ] || exit 2` bit is there to turn the success of `sudo -k true` (which is
-    // expected to fail) into a failure of the entire shell script
-    Command::new("sh")
-        .arg("-c")
-        .arg(format!(
-            "echo {PASSWORD} | sudo -S true; sudo -k true; [ $? -eq 1 ] || exit 2; sudo true"
-        ))
-        .as_user(USERNAME)
-        .exec(&env)?
-        .assert_success()
 }
